+++ src/main/java/org/folio/innreach/domain/service/ContributionCleanupService.java
@@ -1,0 +1,5 @@
+package org.folio.innreach.domain.service;
+
+public interface ContributionCleanupService {
+  void cleanup();
+}